pyproject.toml
[project]
name = "TalentFinder"
version = "1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.132.0",
    "uvicorn>=0.41.0",
    "pydantic>=2.12.5",
    "sqlalchemy==2.0.46",
    "psycopg==3.2.12",
    "pydantic-settings==2.11.0",
    "email-validator==2.3.0",
    "motor==3.7.1",
    "chromadb==1.5.1",
    "langgraph>=0.0.54",
    "langchain>=0.1.0",
    "langchain-openai>=1.1.10",
    "langchain-groq>=0.1.0",
    "tenacity>=8.2.0",
    "python-jose==3.5.0",
    "passlib==1.7.4",
    "python-multipart>=0.0.22",
    "argon2-cffi>=25.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "C4"]
ignore = ["B008", "E501", "B904", "N818"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
ignore_errors = true
strict = false
disallow_untyped_defs = false
explicit_package_bases = true